mv logic determining initial donation status into dedicated class
